code quality maintenance.

+++ services/session/server/avsession_service.cpp
@@ -2496,7 +2496,7 @@ void AVSessionService::NotifySystemUI(const AVSessionDescriptor* historyDescript
         std::make_shared<Notification::NotificationContent>(localLiveViewContent);
     CHECK_AND_RETURN_LOG(content != nullptr, "avsession item notification content nullptr error");
 
-    auto uid = topSession_ ? topSession_->GetUid() : getuid();
+    auto uid = topSession_ ? topSession_->GetUid() : static_cast<int32_t>(getuid());
     request.SetSlotType(Notification::NotificationConstant::SlotType::LIVE_VIEW);
     request.SetNotificationId(0);
     request.SetContent(content);
